[Fix] Fix logging overuse (#12)
[Fix] Fix logging overuse

Currently logging is passed to log file using redirect operator in entrypoint.sh. this creates huge log file which includes info messages.
Furthermore, log file is always open and in case any deletion is made or the file is cleared, logging won't work anymore.
FIX: Create a log file that include warn or greater level messages to log file using a custom logger. max size is set to 5MB and 3 backups at most. debug mode can be enabled normally as it was earlier by using a debug flag with globalmon
MAKE SURE: /var/log/globalmon directory exists and is owned by the user and has permission to write there. It is taken care of in the Dockerfile already.

Reviewed-by: Vladimir Hasko <vladimirhasko@gmail.com>
Reviewed-by: Anton Sidelnikov
Reviewed-by: Muneeb H. Jan <muneebhafeezjan@gmail.com>